Renames file for better readability
