+++ lib/src/every_modifier.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
 import '../due_date.dart';
-import 'constants.dart';
+import 'limited_or_every_handler.dart';
 
 /// An enum that represents the direction of the process inside [EveryModifier].
 /// Used on [EveryModifier.processDate].
